+++ src/main/kotlin/org/example/gomstest/domain/board/data/enums/Major.kt
@@ -1,4 +1,4 @@
-package org.example.gomstest.data.enums
+package org.example.gomstest.domain.board.data.enums
 
 enum class Major {
     BACKEND,
