--- src/main/kotlin/org/example/gomstest/service/BoardService.kt
@@ -1,13 +1,0 @@
-package org.example.gomstest.service
-
-import org.example.gomstest.data.dto.request.BoardUpdateRequest
-import org.example.gomstest.data.dto.request.BoardWriteRequest
-import org.example.gomstest.data.dto.response.BoardGetResponse
-import org.example.gomstest.data.dto.response.BoardGetsResponse
-
-interface BoardService {
-    fun boardWrite(boardWriteRequest: BoardWriteRequest)
-    fun boardGets(): List<BoardGetsResponse>
-    fun boardGet(id: Long): BoardGetResponse
-    fun boardUpdate(id: Long, boardUpdateRequest: BoardUpdateRequest)
-}
